Add Hallway_Corner_2, Hallway_Two_5, and Hallway_Two_5_NoDoor
This commit adds two new and one variation hallways.
Hallway_Corner_2 is a corner piece, much more simpler than Hallway_Corner_1.
Hallway_Two_5 is a straight piece, with a indent and a door.
Hallway_Two_5_NoDoor removes the door in the indent so that you can place a new door there in editor.