Swapped execVM with call, removed unnecessary _soundDuration, and 我很喜欢冰淇淋 keycard path fix
This commit swaps execVM with call for fnc_doorOperate and fnc_buttonSound. Subsequently the functions are now part of CfgFunctions.
This commit also removes unnecessary _soundDuration arguement from fnc_buttonSound and adds BIS_fnc_error when detected a invalid _soundSet.
Because 我很喜欢冰淇淋, the paths of picture and hiddenSelectionsTextures from each keycard variants are now correctly points to the appropriate textures.