v0.5.2 Hotfix - Door with Card Reader script downdated and several Zeus convenience
This commit addresses Door with Card Reader sometimes not work as intended. This reverts the changes made in v0.5 to v0.4.2 for Door with Card Reader. Individual Card Reader is unaffected.
This commit also fixed scopeCurator being curatorScope, removed individual keypads from Zeus, and added new Level 1 to 5 variants for Door with Card Reader only accessible as Zeus, so Zeus has control over the required clearance level of a door.