SCP-1162 Rework
This commit reworks the script of SCP-1162 item exchange.
The changes include the relocation of script from data/SCP-1162 to functions folder.
Changed the search range for scope of a CfgWeapons entry from 0 to 2 so it would only search for items that are available in Arsenal menu... mostly.
As well as the new localisation entry for the action name.
Lastly, the action will now emit a sound depending on if the exchange was successful or the exchange was unsuccessful and the unit's organ has been yoinked by the hole. The damage caused by the organ yoink is ACE3 compatible.

However, using player command as detecting who is using the action might be problematic. Also, canAddItemToVest and rest of them are not reliably detecting whether player can add item or not.